+++ CHANGELOG.md
@@ -38,3 +38,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 * Fixed planning js.
 * Fixed filtering issue with project overview.
 * Changed sprint report form to GET method.
+* Fixed worklog select path.